requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
streamlit
fastapi
uvicorn
requests